Use pipe-open to read from the ->test() command
In order to make verbose=1 give some information about the reports, we
now use a pipe-open to the command so we can evaluate it on the fly.
Some module have so many dependencies that it takes a fair while before
one can print captured output, we want some indication of progress so we
need to process output on the fly.